+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 # Changelog
 
-## 1.0.1
+## [1.2.0] - 2025/01/09
+
+- Added `triangulateFromPointsAndHolePoints` entry point
+- Improvements to internal model representation
+
+## [1.1.0]
+
+- Added ported fix from [mapbox/earcut#91](https://github.com/mapbox/earcut/pull/91)
+
+## [1.0.1]
 
 - Improved 'pana' score
 
-## 1.0.0
+## [1.0.0]
 
 - Initial version